Use a long for PEMethodSymbol.PackedFlags

+++ src/Compilers/Core/Portable/InternalUtilities/EnumUtilties.cs
@@ -56,11 +56,11 @@ namespace Roslyn.Utilities
         }
 
 #if DEBUG
-        internal static bool ContainsAllValues<T>(int mask) where T : struct, Enum, IConvertible
+        internal static bool ContainsAllValues<T>(long mask) where T : struct, Enum, IConvertible
         {
             foreach (T value in GetValues<T>())
             {
-                int val = value.ToInt32(null);
+                long val = value.ToInt64(null);
                 if ((val & mask) != val)
                 {
                     return false;
